Split utilities into multiple scripts.
