+++ src/soep_preparation/clean_existing_variables/pequiv.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from soep_preparation.utilities import (
+from soep_preparation.utilities.series_manipulator import (
     apply_lowest_int_dtype,
     object_to_bool_categorical,
     object_to_int,
