+++ src/soep_preparation/utilities/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities used in various parts of the project."""
